Use all inputs to represent current agent state
Although the assignment suggests that perhaps not all of the information
in the inputs is needed to represent the state, since I don't know at
present what I could safely discard, I would prefer to naïvely keep
everything.

In the future, if it becomes clear which information can be discarded,
that information can be discarded via a subsequent commit.