Run for 1000 trials rather than 100
Currently, the smartcab is ramming into other vehicles.

The hypothesis for this is that it's not getting enough training time
with other vehicles to really understand how it should behave.

Accordingly, increase the amount of time the agent trains, so that it
can encounter other vehicles more, and accordingly learn from those
encounters