Revise answer to optional question
Previously, the calculation for the number of distinct states was incorrect.

Accordingly, revise to (1) be explicit about the calculation, and
(2) use correct maths to arrive at the correct answer.

Additionally, previously the answer opined that certain states were rare,
but seemed to imply that they had a non-zero chance of occurring.

While these rare events have a non-zero chance of occurring _in the wild_,
however, in _this simulation_ many of those events have exactly a 0% chance
of occurring, since other actors are presumed to obey traffic laws.

Accordingly, revise the discussion of those events to explicitly note that
the smartcab won't have an opportunity to see some of the theoretically-
possible states.