Fix state calculation
Previously, the logic had been changed so that the smartcab
moved into a new state, then looked backwards to its previous state
to update its Q-table.

However, this is not the defined implementation of the Q-Learning algorithm.

Accordingly, move into a new state at the beginning of the current iteration,
then peek ahead to the next state, but do not move into it.